renamed metrics package to pdcmetrics because of linter errors
