+++ pkg/pdcmetrics/metrics.go
@@ -1,4 +1,4 @@
-package metrics
+package pdcmetrics
 
 import (
 	"errors"
